+++ SpatialGDK/Source/SpatialGDK/Private/Interop/SpatialStaticComponentView.cpp
@@ -49,6 +49,13 @@ bool USpatialStaticComponentView::HasComponent(Worker_EntityId EntityId, Worker_
 
 void USpatialStaticComponentView::OnAddComponent(const Worker_AddComponentOp& Op)
 {
+	// With dynamic components enabled, it's possible to get duplicate AddComponent ops which need handling idempotently.
+	// For the sake of efficiency we just exit early here.
+	if (HasComponent(Op.entity_id, Op.data.component_id))
+	{
+		return;
+	}
+
 	TUniquePtr<SpatialGDK::Component> Data;
 	switch (Op.data.component_id)
 	{