+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialGameInstance.h
@@ -15,8 +15,9 @@ class USpatialStaticComponentView;
 
 DECLARE_LOG_CATEGORY_EXTERN(LogSpatialGameInstance, Log, All);
 
-DECLARE_EVENT(USpatialGameInstance, FOnConnectedEvent);
-DECLARE_EVENT_OneParam(USpatialGameInstance, FOnConnectionFailedEvent, const FString&);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE(FOnConnectedEvent);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnConnectionFailedEvent, const FString&, Reason);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnPlayerSpawnFailedEvent, const FString&, Reason);
 
 UCLASS(config = Engine)
 class SPATIALGDK_API USpatialGameInstance : public UGameInstance
@@ -53,11 +54,17 @@ public:
 
 	void HandleOnConnected();
 	void HandleOnConnectionFailed(const FString& Reason);
+	void HandleOnPlayerSpawnFailed(const FString& Reason);
 
 	// Invoked when this worker has successfully connected to SpatialOS
-	FOnConnectedEvent OnConnected;
+	UPROPERTY(BlueprintAssignable)
+	FOnConnectedEvent OnSpatialConnected;
 	// Invoked when this worker fails to initiate a connection to SpatialOS
-	FOnConnectionFailedEvent OnConnectionFailed;
+	UPROPERTY(BlueprintAssignable)
+	FOnConnectionFailedEvent OnSpatialConnectionFailed;
+	// Invoked when the player could not be spawned
+	UPROPERTY(BlueprintAssignable)
+	FOnPlayerSpawnFailedEvent OnSpatialPlayerSpawnFailed;
 
 	void SetFirstConnectionToSpatialOSAttempted() { bFirstConnectionToSpatialOSAttempted = true; };
 	bool GetFirstConnectionToSpatialOSAttempted() const { return bFirstConnectionToSpatialOSAttempted; };